# docs/swagger.yaml
basePath: /v1
definitions:
...
  main.LoginPayload:
    properties:
      email:
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
...
  main.UpdateExecPayload:
    properties:
      email:
        type: string
      first_name:
        maxLength: 72
        type: string
      last_name:
        maxLength: 72
        type: string
      role:
        allOf:
        - $ref: '#/definitions/store.Role'
        enum:
        - admin
        - manager
    type: object
...
  store.Exec:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      role:
        $ref: '#/definitions/store.Role'
      updated_at:
        type: string
    type: object
  store.Role:
    enum:
    - admin
    - manager
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleManager
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for ClassNama, a school management platform.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ClassNama
paths:
  /execs:
    get:
      consumes:
      - application/json
      description: Returns a list of all execs
      operationId: getExecs
      produces:
      - application/json
      responses:
        "200":
          description: List of execs
          schema:
            items:
              $ref: '#/definitions/store.Exec'
            type: array
        "500":
          description: Internal server error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get all executives
      tags:
      - Execs
  /execs/{execID}:
    delete:
      consumes:
      - application/json
      description: Deletes an exec by ID
      operationId: deleteExec
      parameters:
      - description: Exec ID
        in: path
        name: execID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Exec not found
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Delete an executive
      tags:
      - Execs
    get:
      consumes:
      - application/json
      description: Returns a single exec by ID (must be set in context via middleware)
      operationId: getExec
      parameters:
      - description: Exec ID
        in: path
        name: execID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Exec object
          schema:
            $ref: '#/definitions/store.Exec'
        "404":
          description: Exec not found
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get a single executive
      tags:
      - Execs
    patch:
      consumes:
      - application/json
      description: Updates an exec. Only non-nil fields in the payload are updated.
        Versioning ensures concurrency safety.
      operationId: updateExec
      parameters:
      - description: Exec ID
        in: path
        name: execID
        required: true
        type: integer
      - description: Exec fields to update
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.UpdateExecPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Updated exec object
          schema:
            $ref: '#/definitions/store.Exec'
        "400":
          description: Bad request / validation failed
          schema: {}
        "404":
          description: Exec not found
          schema: {}
        "409":
          description: Conflict / concurrent update
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Update an executive
      tags:
      - Execs
  /execs/login:
    post:
      consumes:
      - application/json
      description: Login as an Exec (admin or manager) and get a JWT token
      parameters:
      - description: Login payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.LoginPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Returns the logged-in exec and JWT token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Exec Login
      tags:
      - Execs
...
  /students/login:
    post:
      consumes:
      - application/json
      description: Login as a Student and get a JWT token
      parameters:
      - description: Login payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.LoginPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Returns the logged-in student and JWT token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Student Login
      tags:
      - Students
...
  /teachers/login:
    post:
      consumes:
      - application/json
      description: Login as a Teacher and get a JWT token
      parameters:
      - description: Login payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.LoginPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Returns the logged-in teacher and JWT token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Teacher Login
      tags:
      - Teachers
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"